+++ RELEASE_NOTES.md
@@ -1,3 +1,4 @@
+
 <!--
 
     Licensed to the Apache Software Foundation (ASF) under one
@@ -27,11 +28,16 @@
 * [IOTDB-1466] Support device template
 * [IOTDB-1491] UDTF query supported in cluster
 * [IOTDB-1536] Support fuzzy query
+* [IOTDB-1496] Timed flush memtable
 * [IOTDB-1561] Support fill by specific value
+* [IOTDB-1565] Add sql: set system to readonly/writable
+* [IOTDB-1569] Timed close TsFileProcessor
 * TTL can be set to the prefix path of storage group
 * add JMX monitor to all ThreadPools in the server module 
 
 ## Improvements
+* [IOTDB-1566] Do not restrict concurrent write partitions
+* [IOTDB-1585] ModificationFile‘s write interface blocking
 * Use StringCachedPool in TsFileResource to reduce the memory size 
 * write performance optimization when replicaNum == 1
 * Optimize Primitive Array Manager
@@ -57,6 +63,7 @@
 * [IoTDB-1501] Fix compaction recover delete tsfile bug
 * [IOTDB-1529] Fix mlog recover idx bug and synchronize setStorageGroup
 * [IOTDB-1537] fix insertTablet permission
+* [IOTDB-1539] Fix delete operation with value filter is abnormal
 * [IOTDB-1540] Bug Fix: 500 when using IN operator
 * [IOTDB-1541] Fix query result not right due to non-precise time index of resource
 * [IOTDB-1542] Cpp client segment fault: char[] buffer overflow caused by long exception message
@@ -64,6 +71,7 @@
 * [IOTDB-1546] Optimize the Upgrade Tool rewrite logic to reduce the temp memory cost
 * [IOTDB-1552] Only allow equivalent filter for TEXT data type
 * [IOTDB-1556] Abort auto create device when meet exception in setStorageGroup
+* [IOTDB-1574] Deleted file handler leak
 * [ISSUE-3116] Bug when using natural month unit in time interval in group by query
 * [ISSUE-3316] Query result with the same time range is inconsistent in group by query
 * [ISSUE-3436] Fix query result not right after deleting multiple time interval of one timeseries
@@ -71,6 +79,7 @@
 * [ISSUE-3545] Fix Time interval value is disorder in group by month
 * [ISSUE-3653] fix Max_time and last return inconsistent result
 * [ISSUE-3690] Memory leaks on the server when cpp client invokes checkTimeseriesExists
+* [ISSUE-3805] OOM caused by Chunk cache
 *  Fix DataMigrationExample OOM if migrate too many timeseries 
 * Handle false positive cases which may cause NPE of tsfile bloom filter 
 * Fix Windows shell error on JDK11 & fix iotdb-env.bat not working